separate user preview components into their own folder
