+++ src/features/admin/userPreview/SearchForm.jsx
@@ -1,0 +1,29 @@
+import { Search } from "lucide-react"
+
+export default function SearchForm({ searchId, setSearchId, handleSearch }) {
+  return (
+    <div className="bg-white dark:bg-gray-800 rounded-lg shadow-md border border-gray-200 dark:border-gray-700">
+      <div className="p-6">
+        <form onSubmit={handleSearch} className="flex gap-4">
+          <div className="flex-1">
+            <input
+              type="text"
+              placeholder="Enter user ID to search..."
+              value={searchId}
+              onChange={(e) => setSearchId(e.target.value)}
+              className="w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:text-white"
+            />
+          </div>
+          <button
+            type="submit"
+            disabled={!searchId.trim()}
+            className="inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md shadow-sm text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            <Search className="h-4 w-4 mr-2" />
+            Search
+          </button>
+        </form>
+      </div>
+    </div>
+  )
+}
